generified list component
