+++ Client/src/Components/Graph/GraphControls/DatasetControl.tsx
@@ -1,10 +1,10 @@
 import { Button, Grid } from '@material-ui/core'
+import { DatasetRow, IDatasetRowProps } from './DatasetRow'
 
-import { DatasetRow } from '../DatasetList/DatasetRow'
 import { ExportDatasetsButton } from './ExportDatasetsButton'
 import { IDatasetModel } from '../../../Models/Datasets/IDatasetModel'
 import { IGraphDatasetState } from '../../../Models/Graph/IGraphDatasetModel'
-import { PaginatedList } from '../../Utils/PaginatedList'
+import { List } from '../../Utils/List'
 import React from 'react'
 import { SearchViewModal } from '../../Search/SearchViewModal'
 import { toDatasetRows } from '../../../Common/Helpers/GraphHelpers'
@@ -73,11 +73,11 @@ export const DatasetControl = (props: IProps) => {
           </Grid> : null
         }
       </Grid>
-      <PaginatedList
+      <List
         RowComponent={DatasetRow}
-        datasets={toDatasetRows(completeDatasets, datasetStates)}
-        onRemoveDatasetClick={handleDatasetRemoved}
-        onHideDatasetSwitch={onHideDatasetSwitch}
+        models={toDatasetRows(completeDatasets, datasetStates)}
+        rowProps={{ onRemoveDatasetClick: handleDatasetRemoved, onHideDatasetSwitch: onHideDatasetSwitch } as IDatasetRowProps}
+        withPagination
       />
     </>
   )
